+++ CHANGELOG.md
@@ -7,13 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Debug log lines to print payload info when initializing.
+
 ### Changed
+
+- Updated `minecraft` from `1.21` to `1.21.1`
+- Updated `kotlin` from `2.0.0` to `2.0.20`
+- Updated `fabric-api` from `0.100.7+1.21` to `0.103.0+1.21.1`
+- Updated `fabric-kotlin` from `1.11.0+kotlin.2.0.0` to `1.12.1+kotlin.2.0.20`
+- Updated `fabric-loader` from `0.16.0` to `0.16.3`
+- Updated `spotless` from `7.0.0.BETA1` to `7.0.0.BETA2`
+- Moved use of client networking and events to a `client` entrypoint. (TestMod)
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+- Broken custom entrypoint.
 
 ### Security